Recover byte move flags across Player and Baddie

+++ include/main/dll/baddie_setmove.h
@@ -13,7 +13,7 @@
  *
  * Call order is (obj, state, moveId, speed, moveControlFlags, stateByte).
  */
-void baddieSetMove(GameObject* obj, int state, u8 moveId, f32 speed, int moveControlFlags, u8 stateByte);
+void baddieSetMove(GameObject* obj, int state, u8 moveId, f32 speed, u8 moveControlFlags, u8 stateByte);
 #define Baddie_SetMove(obj, state, moveId, speed, moveControlFlags, stateByte)                                        \
     baddieSetMove((GameObject*)(obj), (int)(state), (moveId), (speed), (moveControlFlags), (stateByte))
